requirements.txt
langchain_community
httptools
pinecone-client
supabase
python-dotenv
routes
watchfiles
fastapi
uvicorn
langchain
openai
uvicorn
langchain_openai
langchain_pinecone
python-multipart
pypdf